requirements.txt
torch>=2.0
torchvision
transformers>=4.36
ultralytics>=8.0.0
opencv-python
Pillow
numpy
scikit-learn

